split output file utilities from input data utilities
