+++ 03_run_cancer_type_classification.py
@@ -18,8 +18,9 @@ from pancancer_evaluation.exceptions import (
     OneClassError,
     ResultsFileExistsError
 )
-import pancancer_evaluation.utilities.data_utilities as du
 from pancancer_evaluation.utilities.classify_utilities import run_cv_cancer_type
+import pancancer_evaluation.utilities.data_utilities as du
+import pancancer_evaluation.utilities.file_utilities as fu
 
 def process_args():
     p = argparse.ArgumentParser()
@@ -134,11 +135,11 @@ if __name__ == '__main__':
                 # (or has a different alias, TODO check for this later)
                 print('Gene {} not found in mutation data, skipping'.format(gene),
                       file=sys.stderr)
-                cancer_type_log_df = du.generate_log_df(
+                cancer_type_log_df = fu.generate_log_df(
                     log_columns,
                     [gene, 'N/A', use_pancancer, shuffle_labels, 'gene_not_found']
                 )
-                du.write_log_file(cancer_type_log_df, args.log_file)
+                fu.write_log_file(cancer_type_log_df, args.log_file)
                 continue
 
             inner_progress = tqdm(args.holdout_cancer_types,
@@ -161,7 +162,7 @@ if __name__ == '__main__':
                         print('Skipping because results file exists already: '
                               'gene {}, cancer type {}'.format(gene, cancer_type),
                               file=sys.stderr)
-                    cancer_type_log_df = du.generate_log_df(
+                    cancer_type_log_df = fu.generate_log_df(
                         log_columns,
                         [gene, cancer_type, use_pancancer, shuffle_labels, 'file_exists']
                     )
@@ -170,7 +171,7 @@ if __name__ == '__main__':
                         print('Skipping due to no test samples: gene {}, '
                               'cancer type {}'.format(gene, cancer_type),
                               file=sys.stderr)
-                    cancer_type_log_df = du.generate_log_df(
+                    cancer_type_log_df = fu.generate_log_df(
                         log_columns,
                         [gene, cancer_type, use_pancancer, shuffle_labels, 'no_test_samples']
                     )
@@ -179,13 +180,13 @@ if __name__ == '__main__':
                         print('Skipping due to one holdout class: gene {}, '
                               'cancer type {}'.format(gene, cancer_type),
                               file=sys.stderr)
-                    cancer_type_log_df = du.generate_log_df(
+                    cancer_type_log_df = fu.generate_log_df(
                         log_columns,
                         [gene, cancer_type, use_pancancer, shuffle_labels, 'one_class']
                     )
                 else:
                     # only save results if no exceptions
-                    du.save_results_cancer_type(tcga_data.gene_dir,
+                    fu.save_results_cancer_type(tcga_data.gene_dir,
                                                 tcga_data.check_file,
                                                 results,
                                                 gene,
@@ -193,5 +194,5 @@ if __name__ == '__main__':
                                                 shuffle_labels)
 
                 if cancer_type_log_df is not None:
-                    du.write_log_file(cancer_type_log_df, args.log_file)
+                    fu.write_log_file(cancer_type_log_df, args.log_file)
